[BE] FIX : socket getClothingInfo

+++ BE/smartclothing/src/main/java/sueprtizen/smartclothing/domain/weather/dto/WeatherResponseDTO.java
@@ -2,16 +2,16 @@ package sueprtizen.smartclothing.domain.weather.dto;
 
 public record WeatherResponseDTO(
         int icon,
-        int lowestTemperature,
-        int highestTemperature,
-        int lowestRealFeelingTemperature,
-        int highestRealFeelingTemperature,
-        int precipitation,
-        int snowCover,
-        int humidity,
-        int windSpeed,
-        int solarIrradiance,
-        int UV,
+        double lowestTemperature,
+        double highestTemperature,
+        double lowestRealFeelingTemperature,
+        double highestRealFeelingTemperature,
+        double precipitation,
+        double snowCover,
+        double humidity,
+        double windSpeed,
+        double solarIrradiance,
+        double UV,
         String UVMessage
 ) {
 }
